Finish exercise 04
I got really close on doing this on my own but
couldn't pass one of the tests. I looked at the
solution posted here, and it was interesting
because I originally thought a foreach loop
would work well but I'm not comfortable with
arrow syntax so I opted for a for loop.
I also have never seen ...args. Overall, learned a
lot